+++ dotnet/Data/DriveImportRepository.cs
@@ -397,5 +397,64 @@
 
             response.EnsureSuccessStatusCode();
         }
+
+        public async Task<FolderIds> LoadFolderIds(string accountName)
+        {
+            //Console.WriteLine("-> LoadToken <-");
+            var request = new HttpRequestMessage
+            {
+                Method = HttpMethod.Get,
+                RequestUri = new Uri($"http://infra.io.vtex.com/vbase/v2/{this._httpContextAccessor.HttpContext.Request.Headers[DriveImportConstants.VTEX_ACCOUNT_HEADER_NAME]}/{this._environmentVariableProvider.Workspace}/buckets/{this._applicationName}/{DriveImportConstants.BUCKET}/files/{accountName}")
+            };
+
+            string authToken = this._httpContextAccessor.HttpContext.Request.Headers[DriveImportConstants.HEADER_VTEX_CREDENTIAL];
+            if (authToken != null)
+            {
+                request.Headers.Add(DriveImportConstants.AUTHORIZATION_HEADER_NAME, authToken);
+                request.Headers.Add(DriveImportConstants.VTEX_ID_HEADER_NAME, authToken);
+            }
+
+            request.Headers.Add("Cache-Control", "no-cache");
+
+            var client = _clientFactory.CreateClient();
+            var response = await client.SendAsync(request);
+            string responseContent = await response.Content.ReadAsStringAsync();
+            //Console.WriteLine($"-> LoadToken [{response.StatusCode}] {responseContent} <-");
+
+            if (response.StatusCode == HttpStatusCode.NotFound)
+            {
+                return null;
+            }
+
+            // A helper method is in order for this as it does not return the stack trace etc.
+            response.EnsureSuccessStatusCode();
+
+            FolderIds folderIds = JsonConvert.DeserializeObject<FolderIds>(responseContent);
+
+            return folderIds;
+        }
+
+        public async Task<bool> SaveFolderIds(FolderIds folderIds, string accountName)
+        {
+            var jsonSerializedToken = JsonConvert.SerializeObject(folderIds);
+            var request = new HttpRequestMessage
+            {
+                Method = HttpMethod.Put,
+                RequestUri = new Uri($"http://infra.io.vtex.com/vbase/v2/{this._httpContextAccessor.HttpContext.Request.Headers[DriveImportConstants.VTEX_ACCOUNT_HEADER_NAME]}/{this._environmentVariableProvider.Workspace}/buckets/{this._applicationName}/{DriveImportConstants.BUCKET}/files/{accountName}"),
+                Content = new StringContent(jsonSerializedToken, Encoding.UTF8, DriveImportConstants.APPLICATION_JSON)
+            };
+
+            string authToken = this._httpContextAccessor.HttpContext.Request.Headers[DriveImportConstants.HEADER_VTEX_CREDENTIAL];
+            if (authToken != null)
+            {
+                request.Headers.Add(DriveImportConstants.AUTHORIZATION_HEADER_NAME, authToken);
+                request.Headers.Add(DriveImportConstants.VTEX_ID_HEADER_NAME, authToken);
+            }
+
+            var client = _clientFactory.CreateClient();
+            var response = await client.SendAsync(request);
+
+            return response.IsSuccessStatusCode;
+        }
     }
 }